+++ packages/moonstone/MoonstoneDecorator/MoonstoneDecorator.js
@@ -43,9 +43,13 @@ const defaultConfig = {
  * [floating layer]{@link ui/FloatingLayer.FloatingLayerDecorator},
  * [resolution independence]{@link ui/resolution.ResolutionDecorator},
  * [custom text sizing]{@link moonstone/MoonstoneDecorator.TextSizeDecorator},
- * [spotlight]{@link spotlight.SpotlightRootDecorator}, and
+ * [skin support]{@link ui/Skinnable}, [spotlight]{@link spotlight.SpotlightRootDecorator}, and
  * [internationalization support]{@link i18n/I18nDecorator.I18nDecorator}. It is meant to be applied to
  * the root element of an app.
+ *
+ * [Skins]{@link ui/Skinnable} provide a way to change the coloration of your app. The currently
+ * supported skins for Moonstone are "moonstone" (the default, dark skin) and "moonstone-light".
+ * Use the `skin` property to assign a skin. Ex: `<DecoratedApp skin="moonstone-light" />`
  *
  * @class MoonstoneDecorator
  * @memberof moonstone/MoonstoneDecorator