+++ packages/ui/ViewManager/TransitionGroup.js
@@ -107,6 +107,13 @@ class TransitionGroup extends React.Component {
 		component: PropTypes.any,
 
 		/**
+		 * Current Index the ViewManager is on
+		 *
+		 * @type {Number}
+		 */
+		currentIndex: PropTypes.number,
+
+		/**
 		 * Called when each view is rendered during initial construction.
 		 *
 		 * @type {Function}
@@ -390,9 +397,11 @@ class TransitionGroup extends React.Component {
 		// support wrapping arbitrary children with a component that supports the necessary
 		// lifecycle methods to animate transitions
 		const childrenToRender = this.state.children.map(child => {
+			const isLeaving = child.props['data-index'] !== this.props.currentIndex && typeof child.props['data-index'] !== 'undefined';
+
 			return React.cloneElement(
 				this.props.childFactory(child),
-				{key: child.key, ref: child.key}
+				{key: child.key, ref: child.key, leaving: isLeaving}
 			);
 		});
 
@@ -400,7 +409,9 @@ class TransitionGroup extends React.Component {
 		const props = Object.assign({}, this.props);
 		delete props.size;
 		delete props.childFactory;
+		delete props.currentIndex;
 		delete props.component;
+		delete props.onAppear;
 		delete props.onAppear;
 		delete props.onEnter;
 		delete props.onLeave;