+++ packages/moonstone/IncrementSlider/IncrementSlider.js
@@ -39,6 +39,15 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 
 		propTypes: /** @lends moonstone/IncrementSlider.IncrementSliderBase.prototype */ {
 			/**
+			 * When `true`, the knob displays selected and can be moved using 5-way controls.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			active: PropTypes.bool,
+
+			/**
 			 * Background progress, as a proportion between `0` and `1`.
 			 *
 			 * @type {Number}
@@ -77,6 +86,14 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			incrementIcon: PropTypes.string,
 
 			/**
+			 * The method to run when the input mounts, giving a reference to the DOM.
+			 *
+			 * @type {Function}
+			 * @private
+			 */
+			inputRef: PropTypes.func,
+
+			/**
 			 * The maximum value of the increment slider.
 			 *
 			 * @type {Number}
@@ -93,6 +110,14 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			 * @public
 			 */
 			min: PropTypes.number,
+
+			/**
+			 * The handler when the knob is activated or deactivated by selecting it via 5-way
+			 *
+			 * @type {Function}
+			 * @public
+			 */
+			onActivate: PropTypes.func,
 
 			/**
 			 * The handler to run when the value is changed.
@@ -130,6 +155,42 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			 * @public
 			 */
 			onSpotlightDisappear: PropTypes.func,
+
+			/**
+			 * When `true`, a pressed visual effect is applied
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			pressed: PropTypes.bool,
+
+			/**
+			 * `scrubbing` only has an effect with a detachedKnob, and is a performance optimization
+			 * to not allow re-assignment of the knob's value (and therefore position) during direct
+			 * user interaction.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			scrubbing: PropTypes.bool,
+
+			/**
+			 * The method to run when the slider bar component mounts, giving a reference to the DOM.
+			 *
+			 * @type {Function}
+			 * @private
+			 */
+			sliderBarRef: PropTypes.func,
+
+			/**
+			 * The method to run when mounted, giving a reference to the DOM.
+			 *
+			 * @type {Function}
+			 * @private
+			 */
+			sliderRef: PropTypes.func,
 
 			/**
 			 * When `true`, the component cannot be navigated using spotlight.
@@ -246,42 +307,56 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			incrementAriaLabel: ({['aria-label']: label, value}) => (`${label} ${value} ${$L('press ok button to increase the value')}`)
 		},
 
-		render: ({decrementAriaLabel, decrementDisabled, decrementIcon, incrementAriaLabel, incrementDisabled, incrementIcon, incrementSliderClasses, onIncrement, onDecrement, onSpotlightDisappear, spotlightDisabled, ...rest}) => {
-			delete rest['aria-label'];
-
-			return (
-				<div className={incrementSliderClasses}>
-					<IncrementSliderButton
-						aria-label={decrementAriaLabel}
-						className={css.decrementButton}
-						disabled={decrementDisabled}
-						onClick={onDecrement}
-						onSpotlightDisappear={onSpotlightDisappear}
-						spotlightDisabled={spotlightDisabled}
-					>
-						{decrementIcon}
-					</IncrementSliderButton>
-					<Slider
-						{...rest}
-						className={css.slider}
-						onDecrement={onDecrement}
-						onIncrement={onIncrement}
-						onSpotlightDisappear={onSpotlightDisappear}
-						spotlightDisabled={spotlightDisabled}
-					/>
-					<IncrementSliderButton
-						aria-label={incrementAriaLabel}
-						className={css.incrementButton}
-						disabled={incrementDisabled}
-						onClick={onIncrement}
-						onSpotlightDisappear={onSpotlightDisappear}
-						spotlightDisabled={spotlightDisabled}
-					>
-						{incrementIcon}
-					</IncrementSliderButton>
-				</div>
-			);
-		}
+		render: ({active, backgroundProgress, decrementAriaLabel, decrementDisabled, decrementIcon, disabled, incrementAriaLabel, incrementDisabled, incrementIcon, incrementSliderClasses, inputRef, max, min, onActivate, onChange, onDecrement, onIncrement, onSpotlightDisappear, pressed, scrubbing, sliderBarRef, sliderRef, spotlightDisabled, step, tooltip, tooltipAsPercent, tooltipForceSide, tooltipSide, value, vertical, ...rest}) => (
+			<div {...rest} className={incrementSliderClasses}>
+				<IncrementSliderButton
+					aria-label={decrementAriaLabel}
+					className={css.decrementButton}
+					disabled={decrementDisabled}
+					onClick={onDecrement}
+					onSpotlightDisappear={onSpotlightDisappear}
+					spotlightDisabled={spotlightDisabled}
+				>
+					{decrementIcon}
+				</IncrementSliderButton>
+				<Slider
+					active={active}
+					backgroundProgress={backgroundProgress}
+					className={css.slider}
+					disabled={disabled}
+					inputRef={inputRef}
+					max={max}
+					min={min}
+					onActivate={onActivate}
+					onChange={onChange}
+					onDecrement={onDecrement}
+					onIncrement={onIncrement}
+					onSpotlightDisappear={onSpotlightDisappear}
+					pressed={pressed}
+					scrubbing={scrubbing}
+					sliderBarRef={sliderBarRef}
+					sliderRef={sliderRef}
+					spotlightDisabled={spotlightDisabled}
+					step={step}
+					tooltip={tooltip}
+					tooltipAsPercent={tooltipAsPercent}
+					tooltipForceSide={tooltipForceSide}
+					tooltipSide={tooltipSide}
+					value={value}
+					vertical={vertical}
+				/>
+				<IncrementSliderButton
+					aria-label={incrementAriaLabel}
+					className={css.incrementButton}
+					disabled={incrementDisabled}
+					onClick={onIncrement}
+					onSpotlightDisappear={onSpotlightDisappear}
+					spotlightDisabled={spotlightDisabled}
+				>
+					{incrementIcon}
+				</IncrementSliderButton>
+			</div>
+		)
 	});
 });