+++ packages/moonstone/internal/DateTimeDecorator/DateTimeDecorator.js
@@ -161,6 +161,10 @@ const DateTimeDecorator = hoc((config, Wrapped) => {
 		 * @returns {Number}			Updated internal value
 		 */
 		updateValue = (value) => {
+			const {day, month, year} = value;
+			const maxDays = value.cal.getMonLength(month, year);
+			value.day = (day <= maxDays) ? day : maxDays;
+
 			const date = DateFactory(value);
 			const newValue = date.getTimeExtended();
 			const changed =	this.props.value == null || this.props.value !== newValue;