+++ packages/sampler/stories/moonstone-stories/DatePicker.js
@@ -1,7 +1,7 @@
 import DatePicker, {DatePickerBase} from '@enact/moonstone/DatePicker';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
-import {withKnobs, boolean, text} from '@kadira/storybook-addon-knobs';
+import {boolean, text} from '@kadira/storybook-addon-knobs';
 
 import nullify from '../../src/utils/nullify.js';
 import {mergeComponentMetadata, removeProps} from '../../src/utils/propTables';
@@ -10,7 +10,6 @@ const Config = mergeComponentMetadata('DatePicker', DatePickerBase, DatePicker);
 removeProps(Config, 'year defaultOpen day maxDays maxMonths month onChangeDate onChangeMonth onChangeYear order');
 
 storiesOf('DatePicker')
-	.addDecorator(withKnobs)
 	.addWithInfo(
 		' ',
 		'The basic DatePicker',