+++ packages/ui/CHANGELOG.md
@@ -11,6 +11,11 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Fixed
 
 - `ui/Cancelable` to run modal handlers in the right order
+## Unreleased
+
+### Changed
+
+-`ui/View` to prevent re-renders on View's leaving the ViewManager
 
 ## [1.0.0] - 2017-03-31