+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `ui/Skinnable` to provide themes with a way to apply a base theme styling and skins on top of that
 ## Unreleased
 
 ### Changed