+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,11 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Added
 
 - `ui/Transition` prop `onShow` that fires when transitioning into view a component.
+## Unreleased
+
+### Changed
+
+-`ui/View` to prevent re-renders on views leaving the `ViewManager`
 
 ## [1.1.0] - 2017-04-21