+++ packages/ui/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
-## [unreleased]
+## Unreleased
+
+### Changed
+
+-`ui/View` to prevent re-renders on views leaving the `ViewManager`
+
+## [1.1.0] - 2017-04-21
 
 ### Changed